issue 14 > renamed  functions and created calendar file
